requirements.txt
boto3==1.35.20
botocore==1.35.20
python-dotenv==1.0.1
pytest
